requirements.txt
# Web Framework & Pydantic
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# AI Stack
langchain==0.3.20
langchain-core==0.3.41
langchain-community==0.3.14
# Các gói mở rộng cho LangChain
langchain-openai==0.3.1
langchain-google-genai==2.0.8
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.2
google-generativeai==0.8.3
langchain-tavily==0.2.0

# Database & Auth
sqlalchemy==2.0.36
databases[sqlite]==0.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
firebase-admin==6.6.0

# Data & Tools
pandas==2.2.3
numpy==2.2.1
beautifulsoup4==4.12.3
python-slugify==8.0.4
pillow==11.0.0
psutil==6.1.1
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
